Make Container block a functional component
