+++ src/hoc/index.js
@@ -1,9 +1,11 @@
 import withUniqueId from './withUniqueId';
 import withGridLegacyMigration from './withGridLegacyMigration';
 import withButtonLegacyMigration from './withButtonLegacyMigration';
+import withContainerLegacyMigration from './withContainerLegacyMigration';
 
 export {
 	withUniqueId,
 	withGridLegacyMigration,
 	withButtonLegacyMigration,
+	withContainerLegacyMigration,
 };
